Arduino version
The simulation of the project using Arduino was carried out in Tinkercad. In this simulation, a soil humidity sensor and a DC motor that acts as a fan were included, which is activated when the temperature exceeds 23°C, measured by the TMP36 sensor. This is because the platform did not have the DHT11 or DHT22 sensors, which are commonly used in this type of projects. Additionally, a photoresistor and an LED were used to control the greenhouse lighting. Due to the lack of pumps for the irrigation system, they were simulated using servomotors: one for the irrigation system itself and another for the control of the tank that stores the irrigation water, the level of which is monitored by an ultrasound sensor.
Link to interact with the simulation and see the code for arduino: https://www.tinkercad.com/things/a7wNJ0zapSD-primera-version-para-invernadero