Esp32 version
In this automated greenhouse project using ESP32, sensors and actuators ensure optimal conditions. A DHT22 sensor monitors the temperature and activates a fan if it exceeds 23°C. A soil moisture sensor activates a water pump (servomotor) when the soil is dry. An ultrasonic sensor monitors the irrigation tank's water level and activates a servomotor to fill it if low. Additionally, a photoresistor turns on the lighting when needed.